+++ src/templates/TreeTemplate.php
@@ -44,7 +44,7 @@ $isLiveSite = str_contains(__DIR__, 'midwestmemoriesfamily');
         // This is the treeview component.
         echo '<ul>';
         echo "<li class='folder'><a href='/?i=1' class='path-link'>Home</a></li>";
-        Path::buildTree(IndexGateway::$requestUnixPath);
+        Path::buildTree();
         echo '</ul>';
         ?>
         <div id="debug"><!-- Debug info for iOS goes here as it lacks a real console. --></div>
@@ -62,7 +62,7 @@ $isLiveSite = str_contains(__DIR__, 'midwestmemoriesfamily');
         function setupTemplate() {
             console.log("Updating tree details...");
             const tv = TreeView.init();
-            tv.openLinkInline('<?= $u_linkUrl ?>');
+            tv.openTreeLink('<?= $u_linkUrl ?>');
         }
     </script>
 </div>